pyproject.toml
[project]
name = "agentwebsearch-py"
version = "0.1.0"
description = "Python package for semantic relevant web search results."
readme = "README.md"
authors = [
    { name = "Enrico Goerlitz", email = "rico.goerlitz@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "googlesearch-python>=1.3.0",
    "hnswlib>=0.8.0",
    "langchain>=0.3.26",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "tiktoken>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["src/tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/agentwebsearch"]
